Ref #81 Rename DenovoGeneSetCollection file
